requirements.txt
flake8
aws_cdk.core
aws_cdk.aws-ec2
aws_cdk.aws_iam
aws_cdk.aws_s3
aws_cdk.aws_s3_assets
-e .
